% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-clean.R
\name{add_rta_col}
\alias{add_rta_col}
\title{Add RTAs to trade dataset}
\usage{
add_rta_col(d, path = "mfn-rtas")
}
\arguments{
\item{d}{dataset to add RTAs to}

\item{path}{where to find RTAs dataset}
}
\description{
Add RTAs to trade dataset
}
