% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-predict.R
\name{conciliate_flows}
\alias{conciliate_flows}
\title{Conciliate flows vs mirrored flows}
\usage{
conciliate_flows(
  y,
  subtract_re = FALSE,
  filter_kg = FALSE,
  unit_values = FALSE,
  include_qty = FALSE,
  aggregation_digits = 4,
  replace_unspecified_iso = TRUE,
  path = "../uncomtrade-datasets-arrow/hs-rev2002/parquet/",
  path2 = "../uncomtrade-datasets-arrow/hs-rev2012/parquet/"
)
}
\arguments{
\item{y}{which year to process}

\item{subtract_re}{substract re-imports/exports from data}

\item{filter_kg}{remove observations not expressed in kilograms
(i.e., not available, liters, etc)}

\item{unit_values}{obtain traded unit values (i.e. kilograms per dollar)}

\item{include_qty}{report units in the final dataset (defaults to `FALSE`)}

\item{aggregation_digits}{length of HS codes to use in the data (defaults to `4`)}

\item{replace_unspecified_iso}{convert all ocurrences of `0-unspecified` to
`e-[0-9]` (i.e., `e-439` when corresponding)}

\item{path}{the route to the folder with the parquet files for HS02 data}

\item{path2}{the route to the folder with the parquet files for HS12 data}
}
\description{
Conciliate flows vs mirrored flows
}
