% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-download.R
\name{convert_to_arrow}
\alias{convert_to_arrow}
\title{Convert CSV from UN COMTRADE to Apache Arrow}
\usage{
convert_to_arrow(t, yrs)
}
\arguments{
\item{t}{index COMPLETE...}

\item{yrs}{years COMPLETE...}
}
\description{
COMPLETE DESC...
}
