% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-clean.R
\name{filter_kg}
\alias{filter_kg}
\title{Filter units in kilograms}
\usage{
filter_kg(d, ton = 1000)
}
\arguments{
\item{d}{dataset with traded values}

\item{ton}{how many kilograms is a ton (1000 = BACI, 1016.0469 = UK, 907.18474 = USA)}
}
\description{
Filter units in kilograms
}
