% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-clean.R
\name{open_uncomtrade_arrow}
\alias{open_uncomtrade_arrow}
\title{Add raw UN COMTRADE datasets in Arrow}
\usage{
open_uncomtrade_arrow(path = NULL)
}
\arguments{
\item{path}{the top directory in the Arrow tree}
}
\description{
Add raw UN COMTRADE datasets in Arrow
}
