% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-clean.R
\name{tidy_flows_multiple_codes}
\alias{tidy_flows_multiple_codes}
\title{Tidy flows by completing from older HS/SITC datasets}
\usage{
tidy_flows_multiple_codes(y, path, from, to)
}
\arguments{
\item{y}{year}

\item{path}{the top directory in the Arrow tree}

\item{from}{HS/SITC codes in the data (i.e., "sitc2")}

\item{to}{HS/SITC codes to convert the data to (i.e., "hs92")}
}
\description{
Tidy flows by completing from older HS/SITC datasets
}
