% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-upload.R
\name{update_distances}
\alias{update_distances}
\title{Update distances table}
\usage{
update_distances(con)
}
\arguments{
\item{con}{SQL connection object}
}
\description{
Deletes and uploads the table again
}
