% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-upload.R
\name{update_gdp}
\alias{update_gdp}
\title{Update GDP table}
\usage{
update_gdp(con, path = "attributes")
}
\arguments{
\item{con}{SQL connection object}

\item{path}{directory where the tidy data is}
}
\description{
Deletes and uploads the table again
}
