% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-upload.R
\name{update_gdp_deflator}
\alias{update_gdp_deflator}
\title{Update GDP deflator table}
\usage{
update_gdp_deflator(con, path = "gdp")
}
\arguments{
\item{con}{SQL connection object}

\item{path}{directory where the tidy data is}
}
\description{
Deletes and uploads the table again
}
