% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-upload.R
\name{update_rtas}
\alias{update_rtas}
\title{Update RTAs table}
\usage{
update_rtas(con, path = "rtas")
}
\arguments{
\item{con}{SQL connection object}

\item{path}{directory where the tidy data is}
}
\description{
Deletes and uploads the table again
}
