% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-upload.R
\name{update_tariffs}
\alias{update_tariffs}
\title{Update tariffs table}
\usage{
update_tariffs(con, path = "tariffs")
}
\arguments{
\item{con}{SQL connection object}

\item{path}{directory where the tidy data is}
}
\description{
Deletes and uploads the table again
}
