% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-upload.R
\name{update_vaccine_inputs}
\alias{update_vaccine_inputs}
\title{Update Vaccine Inputs table}
\usage{
update_vaccine_inputs(con)
}
\arguments{
\item{con}{SQL connection object}
}
\description{
Deletes and uploads the table again
}
