% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-upload.R
\name{update_yr}
\alias{update_yr}
\title{Update YR table}
\usage{
update_yr(con, path = "hs-rev2012-tidy")
}
\arguments{
\item{con}{SQL connection object}

\item{path}{directory where the tidy data is}
}
\description{
Deletes and uploads the table again
}
