% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-clean.R
\name{add_gravity_cols}
\alias{add_gravity_cols}
\title{Add bilateral distances to trade dataset}
\usage{
add_gravity_cols(d, path = "attributes")
}
\arguments{
\item{d}{dataset to add distances to}

\item{path}{where to find distances dataset}
}
\description{
Add bilateral distances to trade dataset
}
