% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-clean.R
\name{compute_ratios}
\alias{compute_ratios}
\title{Add CIF/FOB ratios to trade dataset}
\usage{
compute_ratios(d)
}
\arguments{
\item{d}{dataset to add CIF/FOB ratios to}
}
\description{
Add CIF/FOB ratios to trade dataset
}
