% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-download.R
\name{data_downloading}
\alias{data_downloading}
\title{A wrapper to GNU Parallel and wget for faster downloads}
\usage{
data_downloading(download_links)
}
\arguments{
\item{download_link}{A vector of URLs}
}
\description{
COMPLETE DESC...
}
