% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-download.R
\name{data_downloading}
\alias{data_downloading}
\title{A wrapper to GNU Parallel and wget for faster downloads}
\usage{
data_downloading(
  arrow = T,
  token = NULL,
  dataset = NULL,
  remove_old_files = NULL,
  subset_years = NULL
)
}
\arguments{
\item{arrow}{Set to `FALSE` to just download the datasets without
converting to Apache Arrow. Otherwise keep this set to `TRUE`, which
eases posterior analysis and modeling.}
}
\description{
COMPLETE DESC...
}
