% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-clean.R
\name{filter_inter_quantile_range}
\alias{filter_inter_quantile_range}
\title{Filter trade dataset based on IQR}
\usage{
filter_inter_quantile_range(d)
}
\arguments{
\item{d}{dataset to filter}
}
\description{
Filter trade dataset based on IQR
}
