% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-clean.R
\name{tidy_flows}
\alias{tidy_flows}
\title{Reshape UN COMTRADE bulk datasets to put these under tidy data format}
\usage{
tidy_flows(
  year,
  subtract_re = TRUE,
  replace_unspecified_iso = TRUE,
  include_qty = FALSE,
  path = "../uncomtrade-datasets-arrow/hs-rev2002/parquet/",
  path2 = "../uncomtrade-datasets-arrow/hs-rev2012/parquet/"
)
}
\arguments{
\item{year}{which year to reshape}

\item{subtract_re}{subtract re-imports (re-exports) from imports (exports),
when the result is negative, it is converted to 0 (defaults to TRUE)}

\item{replace_unspecified_iso}{convert all ocurrences of `0-unspecified` to
`e-[0-9]` such as `e-439` when corresponding (defaults to FALSE)}

\item{include_qty}{include quantities and their units in the output\
(defaults to FALSE)}

\item{path}{the route to the folder with the parquet files for HS02 data}

\item{path2}{the route to the folder with the parquet files for HS12 data}
}
\description{
This wrapper takes the raw data and re-expresses it under the Tidy Data
logic: exports and imports are on different columns because those have
different units (FOB vs CIF) and each combination
year-origin-destination-product corresponds to a row. This function can also
discount re-imports and re-export and fix missing ISO-3 codes.
}
