% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-clean.R
\name{tidy_flows}
\alias{tidy_flows}
\title{Reshape UN COMTRADE bulk datasets to put these under tidy data format}
\usage{
tidy_flows(
  year,
  subtract_re = TRUE,
  replace_unspecified_iso = TRUE,
  path = "hs-rev2002/parquet",
  path2 = "hs-rev2012/parquet"
)
}
\arguments{
\item{year}{which year to reshape}

\item{subtract_re}{subtract re-imports (re-exports) from imports (exports),
when the result is negative, it is converted to 0}

\item{replace_unspecified_iso}{convert all ocurrences of `0-unspecified` to
`e-[0-9]` (i.e., `e-439` when corresponding)}

\item{path}{the route to the folder with the parquet files for HS02 data}

\item{path2}{the route to the folder with the parquet files for HS12 data}
}
\description{
This wrapper takes the raw data and re-expresses it under the Tidy Data
logic: exports and imports are on different columns because those have
different units (FOB vs CIF) and each combination
year-origin-destination-product corresponds to a row. This function can also
discount re-imports and re-export and fix missing ISO-3 codes.
}
